backend/requirements.txt
redis==6.2.0
python-socketio==5.13.0
uvicorn==0.35.0
pydantic==2.11.7
pytest==8.4.1